moved ServeLoginTemplate to its own file
